+++ data-raw/cetux.r
@@ -9,9 +9,12 @@ cetux <- read.table("data-raw/ipd_recon.dat",header = TRUE) %>%
 agemed <- 57
 agesd <- (83 - 34)/6 # interpret range as 6 SDs. not used
 p_male <- 0.8
+
 # Use mortality rate for median age at randomisation + years of follow up
 # averaged over gender balance at rand
-cetux_bh <- read.table("../paper/guyot_data/mort_usa.txt",header=TRUE,skip=2) %>%
+# current cetux_bh obtained with data from
+# https://mortality.org/Country/Country?cntr=USA downloaded in early 2023
+cetux_bh <- read.table("data-raw/mort_usa.txt",header=TRUE,skip=2) %>%
   filter(Year==1999) %>%
   mutate(Age = as.numeric(ifelse(Age=="110+", "110", Age)),
          Agenext = ifelse(Age==110, Inf, Age+1),
@@ -19,8 +22,9 @@ cetux_bh <- read.table("../paper/guyot_data/mort_usa.txt",header=TRUE,skip=2) %>
          hazard = p_male*Male + (1 - p_male)*Female) %>%
   filter(Trialyear >= 0) %>%
   select(time=Trialyear, hazard)
+use_data(cetux_bh, overwrite=TRUE)
 
-cetux_seer <- read.table("../paper/guyot_data/seer.dat",skip=10,
+cetux_seer <- read.table("data-raw/seer.dat",skip=10,
                          col.names=c("t","num","denom"),nrows = 21) %>%
   mutate(start = t-1) %>%
   select(start,stop=t,r=num,n=denom) %>%
@@ -31,4 +35,3 @@ cetux_seer <- read.table("../paper/guyot_data/seer.dat",skip=10,
 
 use_data(cetux, overwrite=TRUE)
 use_data(cetux_seer, overwrite=TRUE)
-use_data(cetux_bh, overwrite=TRUE)